Change non-breaking space key
Previously the key sequence used to input the non-breaking space was
G-SPC SPC, a sequence which is very easy to press
accidentally. Furthermore, as the non-breaking space looks visually
identical to a regular space, it can be hard to know when this
combination has been hit unintentially. This commit changes the key
combination for the non-breaking space to G-SPC ~ (with the tilde
being derived from the TeX notation for the non-breaking space), which
should hopefully be more resistant to accidental input.