Disambiguate 'dead keys' and 'key sequences'
As well as applying the disambiguation mentioned above, this commit
also updates several tables in the documentation. This is necessary,
as several tables which previously referred to 'dead keys' now need to
refer to 'key sequences'. During this update to the tables, some other
changes were made (e.g. the addition of a new 'Uppercase?' column);
this possibly should have been done as a separate commit, but ended up
being included in this commit instead.